requirements.txt
absl-py==0.15.0
addict==2.4.0
aiohttp==3.8.4
aiohttp-socks==0.8.0
aioimaplib==1.0.1
aiosignal==1.3.1
aiosmtplib==2.0.1
alembic==1.11.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
batchgenerators==0.21
beautifulsoup4==4.11.2
bech32==1.2.0
bleach==6.0.0
blessed==1.19.1
cached-property==1.5.2
cachetools==4.2.4
cattrs==22.2.0
cbor2==5.4.6
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.7
ci-info==0.2.0
click==8.0.3
clinica==0.4.1
clinicadl==1.0.3
cloudpickle==2.2.1
colorama==0.4.4
contextlib2==21.6.0
cpe==1.2.1
cramjam==2.6.2
crcmod==1.7
croniter==1.3.15
cryptography==39.0.2
cycler==0.11.0
databricks-cli==0.17.7
dateutils==0.6.12
DCNv3==1.0
debugpy==1.6.6
decorator==5.1.1
deepdiff==6.3.0
defusedxml==0.7.1
Deprecated==1.2.13
dicom2nifti==2.3.0
docker==6.1.3
docopt==0.6.2
docstring-parser==0.15
dpcpp-cpp-rt==2022.1.0
einops==0.3.2
enlighten==1.10.2
entrypoints==0.4
etelemetry==0.2.2
exceptiongroup==1.1.1
fastapi==0.88.0
fastjsonschema==2.15.3
fastparquet==0.8.1
filelock==3.3.2
Flask==2.2.5
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grabbit==0.2.6
greenlet==2.0.2
grpcio==1.41.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.7.0
html5lib==1.1
huggingface-hub==0.16.4
idna==3.3
imageio==2.9.0
importlib-metadata==4.8.1
importlib-resources==5.12.0
inquirer==2.10.1
intel-cmplr-lib-rt==2022.1.0
intel-cmplr-lic-rt==2022.1.0
intel-opencl-rt==2022.1.0
intel-openmp==2022.1.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isodate==0.6.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jsonargparse==4.22.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.23.6
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keyring==23.4.1
keyrings.alt==3.1
kiwisolver==1.3.2
kornia==0.6.12
lark==1.1.2
lightning-cloud==0.5.37
lightning-utilities==0.9.0
linecache2==1.0.0
lmdb==1.4.1
lxml==4.6.3
Mako==1.2.4
Markdown==3.3.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.4.2
matplotlib-inline==0.1.6
mdurl==0.1.2
MedPy==0.4.0
mistune==2.0.5
mkl==2022.1.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-collections==0.1.0
mlflow==1.30.1
mmaction==0.5.0
mmcls==0.25.0
mmcv==1.5.1
mmcv-full==1.5.0
mmdet==2.28.1
mmsegmentation==0.27.0
model-index==0.1.11
monai==0.7.0
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.6.3
nibabel==3.2.1
niflow-nipype1-workflows==0.0.5
nilearn==0.8.1
nipype==1.7.0
notebook==6.5.3
notebook_shim==0.2.2
num2words==0.5.10
numpy==1.19.5
nvidia-ml-py==11.525.131
nvitop==1.1.2
oauthlib==3.2.2
olefile==0.46
opencv-contrib-python==4.5.4.58
opencv-python==4.5.2.52
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.10
ordered-set==4.1.0
oss2==2.17.0
packaging==21.2
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.0
prefixed==0.3.2
prettytable==3.7.0
prometheus-client==0.16.0
prometheus-flask-exporter==0.22.4
prompt-toolkit==3.0.38
protobuf==3.19.1
prov==2.0.0
psutil==5.9.4
ptyprocess==0.7.0
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybids==0.5.1
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.11
pyDeprecate==0.3.2
pydicom==2.1.2
pydot==1.4.2
Pygments==2.14.0
PyJWT==2.7.0
pynvml==11.0.0
pyOpenSSL==23.1.1
pyparsing==2.4.7
pyrsistent==0.19.3
python-bitcoinlib==0.11.0
python-dateutil==2.8.2
python-editor==1.0.4
python-gdcm==3.0.21
python-multipart==0.0.6
python-snappy==0.6.1
python-socks==2.2.0
pytorch-lightning==1.8.4
pytz==2023.3.post1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==25.0.1
qtconsole==5.4.1
QtPy==2.3.0
querystring-parser==1.2.4
rdflib==6.0.2
readchar==4.0.5
regex==2022.10.31
requests==2.28.2
requests-cache==1.0.1
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rsa==4.7.2
scalecodec==1.0.37
scikit-image==0.18.1
scikit-learn==0.23.0
scipy==1.2.3
SecretStorage==3.3.3
Send2Trash==1.8.0
SimpleITK==2.1.1.2
simplejson==3.17.5
six==1.16.0
sklearn==0.0.post4
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.48
sqlparse==0.4.4
starlette==0.22.0
starsessions==1.3.0
stix2-patterns==2.0.0
stix2-validator==3.1.3
synapse==2.130.2
synapseclient==2.7.1
tabulate==0.9.0
tbb==2021.6.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.6
termcolor==2.3.0
terminado==0.17.1
terminaltables==3.1.10
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.6.11
tinycss2==1.2.1
toml==0.10.2
torch==1.10.0
torchaudio==0.10.0
torchmetrics==0.11.1
torchvision==0.11.1
tornado==6.2
tqdm==4.65.2
traceback2==1.4.0
traitlets==5.9.0
traits==6.3.1
typeshed-client==2.3.0
typing_extensions==4.5.0
unittest2==1.1.0
uri-template==1.2.0
url-normalize==1.4.3
urllib3==1.26.7
uvicorn==0.22.0
vcrpy==4.1.1
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.15.0
xgboost==0.80
xlrd==1.2.0
xvfbwrapper==0.2.9
xxhash==3.2.0
yapf==0.31.0
yarl==1.9.2
zipp==3.6.0
